pyproject.toml
[project]
name = "taskweaverremake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "jsonschema>=4.23.0",
    "jupyter-client>=8.6.3",
    "matplotlib>=3.10.0",
    "nbformat>=5.10.4",
    "openai>=1.62.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "tenacity>=9.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
